Move interfaces to own files.
