+++ UploadDaemon/Archiving/ArchiveFactory.cs
@@ -3,17 +3,6 @@
 namespace UploadDaemon.Archiving
 {
     /// <summary>
-    /// A factory for <see cref="IArchive"/>s.
-    /// </summary>
-    public interface IArchiveFactory
-    {
-        /// <summary>
-        /// Creates an archive based in the given directory.
-        /// </summary>
-        IArchive CreateArchive(string baseDirectoryPath);
-    }
-
-    /// <summary>
     /// A factory creating <see cref="Archive"/>s.
     /// </summary>
     public class ArchiveFactory : IArchiveFactory
