+++ UploadDaemon/Archiving/DateTimeProvider.cs
@@ -3,17 +3,6 @@
 namespace UploadDaemon.Archiving
 {
     /// <summary>
-    /// A provider that supplies the current date time.
-    /// </summary>
-    public interface IDateTimeProvider
-    {
-        /// <summary>
-        /// Current date time.
-        /// </summary>
-        DateTime Now { get; }
-    }
-
-    /// <summary>
     /// DateTime provider supplying DateTime.Now.
     /// </summary>
     public class DefaultDateTimeProvider : IDateTimeProvider
