+++ UploadDaemon/UploadDaemon.csproj
@@ -99,6 +99,9 @@
     <Compile Include="Archiving\Archive.cs" />
     <Compile Include="Archiving\ArchiveFactory.cs" />
     <Compile Include="Archiving\DateTimeProvider.cs" />
+    <Compile Include="Archiving\IArchive.cs" />
+    <Compile Include="Archiving\IArchiveFactory.cs" />
+    <Compile Include="Archiving\IDateTimeProvider.cs" />
     <Compile Include="MessageFormatter.cs" />
     <Compile Include="Archiving\PurgeArchiveTask.cs" />
     <Compile Include="SymbolAnalysis\ILineCoverageSynthesizer.cs" />
